requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
sentence-transformers==2.2.2
chromadb==0.4.15
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0
torch>=2.2.0
transformers==4.34.0
numpy>=1.26.0
scipy>=1.11.3
supabase==2.3.4